requirements.txt
Flask==1.1.2
pip~=20.2.3
py~=1.9.0
colorama~=0.4.3
Jinja2~=2.11.2
Werkzeug~=1.0.1
itsdangerous~=1.1.0
click~=7.1.2
python-dotenv~=0.14.0
MarkupSafe~=1.1.1
setuptools~=50.3.0
flask-wtf==0.14.3
